a readable solution to a IQ test about free beer
Try to find the answer of a IQ test by writing a simple Go program.

Try paste the code into https://play.golang.org
and click “Run”

try to be simple, so that non-programmer can still able to read and
able to understand how do we ask computer to help solve a problem.